Core: slot map and byte stream (m0-foundation/T9, T10); fix the narrowing CI caught in T7
SlotMap<T> over a std::vector with generation-checked handles, a
LIFO free list, slot-order iteration that identical histories
reproduce, and a sorted key index for FindByKey and the ascending-key
traversal the state hash and the snapshot use. ByteWriter and
ByteReader are the bounded, versioned stream: little-endian integers
written a byte at a time, length-prefixed spans with a caller's
limit, a magic-and-version header with a range check, and a reader
that stays failed. Tests cover reuse, stale handles, the generation
wrap rule, identical histories, ascending keys after out-of-order
erases, every width round-tripped and pinned byte for byte,
truncation at every byte, hostile lengths and the header.

BinaryAngle.cpp initialised its table mask from size() without a
cast; MSVC's C4267 under /WX failed the build, and the cast is added.

+++ Core/BinaryAngle.cpp
@@ -13,7 +13,7 @@ namespace
 // low six bits interpolate toward the next; the interpolation is linear in integers and exact
 // for what it is, with a maximum error against a real sine that the test pins.
 constexpr int TABLE_SHIFT = 6;
-constexpr std::uint32_t TABLE_MASK = SIN_TABLE_16_16.size() - 1;
+constexpr std::uint32_t TABLE_MASK = static_cast<std::uint32_t>(SIN_TABLE_16_16.size() - 1);
 constexpr std::int32_t FRACTION_MASK = (1 << TABLE_SHIFT) - 1;
 
 } // namespace
